minor refactor, errors moved to models
