+++ lib/browser/mixins/browser_info.ts
@@ -1,6 +1,6 @@
 import BrowserClick from './browser_click';
 
-import { QueryError } from '../../errors';
+import { QueryError } from '../../models/errors';
 import { WendigoSelector, GeoLocationCoords } from '../../types';
 import FailIfNotLoaded from '../../decorators/fail_if_not_loaded';
 import { PDFOptions } from '../../puppeteer_wrapper/puppeteer_types';
