+++ lib/models/errors.ts
